+++ app/lizard-nxt-filters.js
@@ -527,8 +527,13 @@ angular.module('lizard-nxt-filters')
 angular.module('lizard-nxt-filters')
   .filter('discreteRasterType', function () {
   return function (input) {
-    var labelParts = input.split(' - ');
-    return labelParts[labelParts.length - 1];
+    if (input) {
+      var labelParts = input.split(' - ');
+      return labelParts[labelParts.length - 1];
+    }
+    else {
+      return '';
+    }
   };
 });
 
@@ -540,10 +545,15 @@ angular.module('lizard-nxt-filters')
 angular.module('lizard-nxt-filters')
   .filter('discreteRasterSource', function () {
   return function (input) {
+    if (input) {
     var labelParts = input.split(' - ');
     var output = labelParts[labelParts.length - 2];
     if (output === undefined) { output = ''; }
     return output;
+    }
+    else {
+      return '';
+    }
   };
 });